Zip.Compress.Deflate: Allocate buffers on heap
When using the deflate compression method in a multi-tasking application
the application has been terminated with a STORAGE_ERROR exception. This
was caused by the static file buffer allocations for the input and output
buffer which are just too large for the default stack size when used in
a tasking environment.

Allocating the input and output buffers on the heap (as it is done for
the Shrink method already) solved the issue.

Also fix a memory leak concerning the disposal of the already heap allo-
cated 'lz_buffer': in the event of the Compression_unefficient exception
the allocated buffer was not disposed correctly.